Implement STAP-A aggregation and FU-A payload splitting

+++ Sources/RTP/Types.swift
@@ -1,8 +1,8 @@
 import Foundation
 
 protocol Payloader {
-    var mtu: UInt64 { get }
-    
+    var mtu: UInt16 { get }
+
     func payload(_ data: Data) -> [Data]
 }
